# perf: use single zstd decoder