--- a/route/otlp_logs_test.go
+++ b/route/otlp_logs_test.go
@@ -46,7 +46,12 @@ func TestLogsOTLPHandler(t *testing.T) {
 	mockTransmission.Start()
 	defer mockTransmission.Stop()
 	mockCollector := collect.NewMockCollector()
-	decoders, err := makeDecoders(1)
+	zstdDecoder, err := zstd.NewReader(
+		nil,
+		zstd.WithDecoderConcurrency(1),
+		zstd.WithDecoderLowmem(true),
+		zstd.WithDecoderMaxMemory(8*1024*1024),
+	)
 	if err != nil {
 		t.Error(err)
 	}
@@ -62,7 +67,7 @@ func TestLogsOTLPHandler(t *testing.T) {
 			incomingOrPeer: "incoming",
 		},
 		Logger:           logger,
-		zstdDecoders:     decoders,
+		zstdDecoder:      zstdDecoder,
 		environmentCache: newEnvironmentCache(time.Second, nil),
 		Sharder: &sharder.SingleServerSharder{
 			Logger: logger,