--- a/route/otlp_trace_test.go
+++ b/route/otlp_trace_test.go
@@ -89,7 +89,12 @@ func TestOTLPHandler(t *testing.T) {
 	mockTransmission := &transmit.MockTransmission{}
 	mockTransmission.Start()
 	defer mockTransmission.Stop()
-	decoders, err := makeDecoders(1)
+	zstdDecoder, err := zstd.NewReader(
+		nil,
+		zstd.WithDecoderConcurrency(1),
+		zstd.WithDecoderLowmem(true),
+		zstd.WithDecoderMaxMemory(8*1024*1024),
+	)
 	if err != nil {
 		t.Error(err)
 	}
@@ -118,7 +123,7 @@ func TestOTLPHandler(t *testing.T) {
 			incomingOrPeer: "incoming",
 		},
 		Logger:           &logger.MockLogger{},
-		zstdDecoders:     decoders,
+		zstdDecoder:      zstdDecoder,
 		environmentCache: newEnvironmentCache(time.Second, nil),
 		Tracer:           noop.Tracer{},
 	}