--- a/route/route.go
+++ b/route/route.go
@@ -52,10 +52,6 @@ import (
 )
 
 const (
-	// numZstdDecoders is set statically here - we may make it into a config option
-	// A normal practice might be to use some multiple of the CPUs, but that goes south
-	// in kubernetes
-	numZstdDecoders        = 4
 	traceIDShortLength     = 8
 	traceIDLongLength      = 16
 	GRPCMessageSizeMax int = 5_000_000 // 5MB
@@ -88,7 +84,7 @@ type Router struct {
 	// iopLogger is a logger that knows whether it's incoming or peer
 	iopLogger iopLogger
 
-	zstdDecoders chan *zstd.Decoder
+	zstdDecoder *zstd.Decoder
 
 	server     *http.Server
 	grpcServer *grpc.Server
@@ -231,9 +227,9 @@ func (r *Router) LnS() {
 	r.environmentCache = newEnvironmentCache(r.Config.GetEnvironmentCacheTTL(), r.lookupEnvironment)
 
 	var err error
-	r.zstdDecoders, err = makeDecoders(numZstdDecoders)
+	r.zstdDecoder, err = makeDecoders(0)
 	if err != nil {
-		r.iopLogger.Error().Logf("couldn't start zstd decoders: %s", err.Error())
+		r.iopLogger.Error().Logf("couldn't start zstd decoder: %s", err.Error())
 		return
 	}
 
@@ -830,7 +826,8 @@ func (r *Router) readAndCloseMaybeCompressedBody(req *http.Request) (*bytes.Buff
 	defer req.Body.Close()
 
 	var reader io.Reader
-	switch req.Header.Get("Content-Encoding") {
+	encodingScheme := req.Header.Get("Content-Encoding")
+	switch encodingScheme {
 	case "gzip":
 		gzipReader, err := gzip.NewReader(req.Body)
 		if err != nil {
@@ -838,18 +835,6 @@ func (r *Router) readAndCloseMaybeCompressedBody(req *http.Request) (*bytes.Buff
 		}
 		defer gzipReader.Close()
 		reader = gzipReader
-	case "zstd":
-		zReader := <-r.zstdDecoders
-		defer func(zReader *zstd.Decoder) {
-			zReader.Reset(nil)
-			r.zstdDecoders <- zReader
-		}(zReader)
-
-		err := zReader.Reset(req.Body)
-		if err != nil {
-			return nil, err
-		}
-		reader = zReader
 	default:
 		reader = req.Body
 	}
@@ -858,6 +843,21 @@ func (r *Router) readAndCloseMaybeCompressedBody(req *http.Request) (*bytes.Buff
 	if _, err := io.Copy(buf, io.LimitReader(reader, HTTPMessageSizeMax)); err != nil {
 		recycleHTTPBodyBuffer(buf)
 		return nil, err
+	}
+
+	if encodingScheme == "zstd" {
+		decompressedBuf := httpBodyBufferPool.Get().(*bytes.Buffer)
+		decompressed, err := r.zstdDecoder.DecodeAll(buf.Bytes(), decompressedBuf.Bytes())
+		recycleHTTPBodyBuffer(buf)
+		if err != nil {
+			recycleHTTPBodyBuffer(decompressedBuf)
+			return nil, err
+		}
+
+		// Update buffer with decompressed data
+		decompressedBuf.Reset()
+		decompressedBuf.Write(decompressed)
+		return decompressedBuf, nil
 	}
 
 	return buf, nil
@@ -907,21 +907,19 @@ func getEventTime(etHeader string) time.Time {
 	return eventTime.UTC()
 }
 
-func makeDecoders(num int) (chan *zstd.Decoder, error) {
-	zstdDecoders := make(chan *zstd.Decoder, num)
-	for i := 0; i < num; i++ {
-		zReader, err := zstd.NewReader(
-			nil,
-			zstd.WithDecoderConcurrency(1),
-			zstd.WithDecoderLowmem(true),
-			zstd.WithDecoderMaxMemory(8*1024*1024),
-		)
-		if err != nil {
-			return nil, err
-		}
-		zstdDecoders <- zReader
+func makeDecoders(concurrency int) (*zstd.Decoder, error) {
+	d, err := zstd.NewReader(
+		nil,
+		// When a value of 0 is provided GOMAXPROCS will be used.
+		// By default this will be set to 4 or GOMAXPROCS, whatever is lower.
+		zstd.WithDecoderConcurrency(concurrency),
+		zstd.WithDecoderLowmem(true),
+		zstd.WithDecoderMaxMemory(8*1024*1024),
+	)
+	if err != nil {
+		return nil, err
 	}
-	return zstdDecoders, nil
+	return d, nil
 }
 
 func unmarshal(r *http.Request, data []byte, v interface{}) error {