--- a/route/route_test.go
+++ b/route/route_test.go
@@ -1345,7 +1345,6 @@ func BenchmarkRouterBatch(b *testing.B) {
 	b.Run("batch_msgpack", func(b *testing.B) {
 		batchMsgpack, err := msgpack.Marshal(batchEvents.events)
 		require.NoError(b, err)
-		body := compressData(batchMsgpack)
 
 		// Create HTTP request directly without server
 		req, err := http.NewRequest("POST", "/1/batch/test-dataset", nil)
@@ -1354,7 +1353,6 @@ func BenchmarkRouterBatch(b *testing.B) {
 		}
 		req.Header.Set("Content-Type", "application/x-msgpack")
 		req.Header.Set("X-Honeycomb-Team", "test-api-key")
-		req.Header.Set("Content-Encoding", "zstd")
 
 		// Set up mux variables for dataset extraction
 		req = mux.SetURLVars(req, map[string]string{"datasetName": "test-dataset"})
@@ -1364,7 +1362,7 @@ func BenchmarkRouterBatch(b *testing.B) {
 
 		b.ResetTimer()
 		for b.Loop() {
-			req.Body = io.NopCloser(bytes.NewReader(body))
+			req.Body = io.NopCloser(bytes.NewReader(batchMsgpack))
 
 			router.batch(w, req)